Refactored dropdown page
