+++ src/test/java/com/epam/tc/hw3/pages/PageObjectDifferentElements.java
@@ -132,29 +132,4 @@ public class PageObjectDifferentElements {
         softRadioLog.assertAll();
     }
 
-    public void selectDropDown(String color) {
-        PageObjectDifferentElements pageObjectDifferentElements = new PageObjectDifferentElements(webDriver);
-        pageObjectDifferentElements.getDropDownComponent().clickOnDropDown();
-        Select dropDownColors = new Select(pageObjectDifferentElements.getDropDownComponent().getDropDownColor());
-        dropDownColors.selectByVisibleText(color);
-    }
-
-    public void assertDropDown(String color) {
-        PageObjectDifferentElements pageObjectDifferentElements = new PageObjectDifferentElements(webDriver);
-        pageObjectDifferentElements.getDropDownComponent().clickOnDropDown();
-        Select dropDownColors = new Select(pageObjectDifferentElements.getDropDownComponent().getDropDownColor());
-        String selectedOption = dropDownColors.getFirstSelectedOption().getText();
-        SoftAssertions softDropDown = new SoftAssertions();
-        softDropDown.assertThat(selectedOption).isEqualTo(color);
-        softDropDown.assertAll();
-    }
-
-    public void assertDropDownLogger(String color) {
-        By loggerPath = new By.ByXPath("//*[contains(text(),'Colors: value changed to " + color + "')]");
-        WebElement colorLoggerElement = webDriver.findElement(loggerPath);
-        SoftAssertions softDropDownLog = new SoftAssertions();
-        softDropDownLog.assertThat(colorLoggerElement.isDisplayed());
-        softDropDownLog.assertAll();
-    }
-
 }
